Remove unused code around grpc server.
